Split DeleteStatementContext and UpdateStatementContext to base context
